Fix laser and clear code

+++ MK4duo/src/planner/planner.cpp
@@ -880,21 +880,20 @@ void Planner::_buffer_line(const float &a, const float &b, const float &c, const
     // interval between steps for X, Y, Z, E, L to feed to the motion control code.
     if (laser.mode == RASTER || laser.mode == PULSED) {
       block->steps_l = labs(block->millimeters * laser.ppm);
-      for (int i = 0; i < LASER_MAX_RASTER_LINE; i++) {
+      for (uint8_t i = 0; i < LASER_MAX_RASTER_LINE; i++) {
         // Scale the image intensity based on the raster power.
         // 100% power on a pixel basis is 255, convert back to 255 = 100.
         #if ENABLED(LASER_REMAP_INTENSITY)
           const int NewRange = (laser.rasterlaserpower * 255.0 / 100.0 - LASER_REMAP_INTENSITY);
-          float NewValue = (float)(((((float)laser.raster_data[i] - 0) * NewRange) / 255.0) + LASER_REMAP_INTENSITY);
+          float     NewValue = (float)(((((float)laser.raster_data[i] - 0) * NewRange) / 255.0) + LASER_REMAP_INTENSITY);
         #else
-		  const int NewRange = (laser.rasterlaserpower * 255.0 / 100.0);
-          float NewValue = (float)(((((float)laser.raster_data[i] - 0) * NewRange) / 255.0));
+          const int NewRange = (laser.rasterlaserpower * 255.0 / 100.0);
+          float     NewValue = (float)(((((float)laser.raster_data[i] - 0) * NewRange) / 255.0));
         #endif
 
         #if ENABLED(LASER_REMAP_INTENSITY)
-        // If less than 7%, turn off the laser tube.
-        if (NewValue <= LASER_REMAP_INTENSITY)
-          NewValue = 0;
+          // If less than 7%, turn off the laser tube.
+          if (NewValue <= LASER_REMAP_INTENSITY) NewValue = 0;
         #endif
 
         block->laser_raster_data[i] = NewValue;
